+++ HISTORY.md
@@ -1,3 +1,9 @@
+0.1.6 / 13.09.2016
+==================
+
+  * Added `scroll` helper
+  * Now autofocusing invalid fields and fields with errors (and scrolling to them too)
+
 0.1.2 / 05.09.2016
 ==================